+++ qml/RoomDelegate.qml
@@ -1,7 +1,6 @@
 import QtQuick 2.5
 import QtQuick.Controls 2.15
 import QtQuick.Layouts 1.3
-import QtQuick.Dialogs 1.2
 import GlobalObject 1.0
 import MatrixClient 1.0
 import Rooms 1.0