+++ qml/RoomList.qml
@@ -31,6 +31,8 @@ Page {
         id: directChatFactory
         AddUserDialog {
             title: "Direct Chat"
+            x: (qmlLibRoot.width - width) / 2
+            y: (qmlLibRoot.height - height) / 2
         }
     }